Refactor to common test beans
